refactor: UserDetail 페이지 폴더에서 Page 접미사 제거
