+++ apps/admin/src/pages/UserDetail/UserDetail.css.ts
